restructured ui. Probably need to do another restructure to put a 'search' or 'query' button in one common spot.
Also we are now popping a new window up when a pattern instance is selected. The plan is for this window to hold class and sequencew diagram images and/or textual representations of the pattern instance. Also, this window will have the evolution of the pattern instance, in sequential form.