Some refactorings and added comments
starting building data structures for sequence diagram stuff. Its going to be a trickier problem than I initially foresaw because of the use of external libraries.. When I track a call chain through a project I need to match the call to the owning class/interface/enum. Its easy if its within the same class as the call.. maybe thats a good startin gpoint.