Finished with selenium analysis. Onto other projects.
Fixed a bug where the pattern detection tool output says the generic type holder (List, for example) is the acutal data type.. This is not true and its a feature of the pattern4 tool because of how generics are handled by java bytecode. So I just put in a greedy search when matching uml operations to check names if all else fails.

Also, did some outputter reformatting.