Fixed a bug with calls in cases of unnecessary actions-type grime.
I noticed that some cases of repetition grime are not working, specifically when a non-pattern classes uses a pattern class... hmm is this really grime in that case? idk.. Its so hard to tell when a client class wants to use a pattern class, and if that client class uses the pattern class more times than once should I consider it bad?